Wait for iOS simulator before running tests
